refactor: formatter.ts 함수 추가
