+++ hanarowa/src/app/mypage/info/page.tsx
@@ -9,30 +9,15 @@ import {
   Modal,
   DatePicker,
 } from '@/components';
+import {
+  formatDateFromISO,
+  formatDateToISO,
+  formatPhone,
+} from '@/utils/formatter';
 import { useGetMemberInfo, useModifyInfo } from '@apis';
 import { useModal } from '@hooks';
 import { useRouter } from 'next/navigation';
 import { useState, ChangeEvent } from 'react';
-
-const digits = (s: string) => s.replace(/\D/g, '');
-
-const formatPhone = (v: string) => {
-  const d = digits(v).slice(0, 11);
-  if (d.length <= 3) return d;
-  if (d.length <= 7) return `${d.slice(0, 3)}-${d.slice(3)}`;
-  return `${d.slice(0, 3)}-${d.slice(3, 7)}-${d.slice(7, 11)}`;
-};
-
-// YYYYMMDD -> YYYY-MM-DD 변환
-const formatDateToISO = (dateStr: string) => {
-  if (!dateStr || dateStr.length !== 8) return '';
-  return `${dateStr.slice(0, 4)}-${dateStr.slice(4, 6)}-${dateStr.slice(6, 8)}`;
-};
-
-// YYYY-MM-DD -> YYYYMMDD 변환
-const formatDateFromISO = (isoDate: string) => {
-  return isoDate.replace(/\D/g, '');
-};
 
 const Page = () => {
   // 서버에서 가져온 값
